Update to fix shortcodes
I think this should make the shortcodes work again, they changed some of the logic in WP 2.5 and the syntax of the shortcode_atts(). 

Also, I changed the function names for good measure so they're definitely the ones being called as they'd already been declared globally so the reference to the plugin object shouldn't be required if the function name is unique.

Lastly just removed the template reference all the way through the chain, there's only one template to call and it's was already defined as the default (now it's just defined in the final function) so it seemed redundant, to make it the default and pass it in the shortcode and the intermediate functions.

I can't test as i've not got an environment that doesn't already have a bunch of complex workarounds it so i've just had to stick to the codex and go off pervious experience. If you want to kick it back that's understandable but if you can try it and it works, grand, if not I'll set up a test environment for this when I've got some more time